- Aggiunti permessi a "business" riguardanti "Gestione File/Cartelle";

+++ app/Http/Livewire/FolderFileManagement/Show.php
@@ -5,11 +5,14 @@
 	use App\File as FileModel;
 	use App\Folder;
 	use App\Folder as FolderModel;
+	use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 	use Illuminate\Support\Facades\Storage;
 	use LivewireUI\Modal\ModalComponent;
 
 	class Show extends ModalComponent
 	{
+		use AuthorizesRequests;
+
 		public FolderModel $folder;
 
 		protected $listeners = [
@@ -22,10 +25,12 @@
 		}
 
 		public function download(FileModel $file) {
+			$this->authorize('download_folder_files');
 			return Storage::download($file->file);
 		}
 
 		public function delete(FileModel $file) {
+			$this->authorize('delete_folder_files');
 			$file->delete();
 			Storage::delete($file->file);
