+++ resources/views/livewire/folder-file-management/index.blade.php
@@ -1,12 +1,14 @@
 <x-slot name="header">
 	<x-page-header>
 		Gestione File e Cartelle
-		<x-slot name="actions">
-			<x-button prepend="plus" iconColor="text-white"
-			          x-on:click="Livewire.emit('openModal', 'folder-file-management.create-folder')">
-				Aggiungi Cartella
-			</x-button>
-		</x-slot>
+		@can('create_folders')
+			<x-slot name="actions">
+				<x-button prepend="plus" iconColor="text-white"
+				          x-on:click="Livewire.emit('openModal', 'folder-file-management.create-folder')">
+					Aggiungi Cartella
+				</x-button>
+			</x-slot>
+		@endcan
 	</x-page-header>
 </x-slot>
 <x-card>
@@ -29,28 +31,34 @@
 					<x-table.td>{{$folder->user->user_data->name}}</x-table.td>
 					<x-table.td>
 						<div class="flex items-center space-x-3">
-							<x-icon name="eye" class="w-5 h-5 text-indigo-500 flex-shrink-0"
-							        wire:click="$emit('openModal', 'folder-file-management.show', {{ json_encode([$folder->id]) }})"></x-icon>
-							<x-icon name="pencil-alt" class="w-5 h-5 text-indigo-500 flex-shrink-0"
-							        wire:click="$emit('openModal', 'folder-file-management.edit-folder', {{ json_encode([$folder->id]) }})"></x-icon>
-							<x-modal>
-								<x-slot name="trigger">
-									<div class="text-red-600 hover:text-red-900">
-										<x-icon name="trash" class="w-5 h-5"></x-icon>
-									</div>
-								</x-slot>
-								<x-slot name="title">
-									Conferma eliminazione
-								</x-slot>
-								Sei sicuro di voler eliminare la cartella "{{ $folder->name }}"?
-								<x-slot name="footer">
-									<x-link-button x-on:click="open = false">Annulla</x-link-button>
-									<x-danger-button class="ml-2" wire:click="deleteFolder({{ $folder->id }})"
-									                 wire:loading.attr="disabled">
-										Elimina
-									</x-danger-button>
-								</x-slot>
-							</x-modal>
+							@can('read_folders')
+								<x-icon name="eye" class="w-5 h-5 text-indigo-600 hover:text-indigo-900 cursor-pointer flex-shrink-0"
+								        wire:click="$emit('openModal', 'folder-file-management.show', {{ json_encode([$folder->id]) }})"></x-icon>
+							@endcan
+							@can('update_folders')
+								<x-icon name="pencil-alt" class="w-5 h-5 text-indigo-600 hover:text-indigo-900 cursor-pointer flex-shrink-0"
+								        wire:click="$emit('openModal', 'folder-file-management.edit-folder', {{ json_encode([$folder->id]) }})"></x-icon>
+							@endcan
+							@can('delete_folders')
+								<x-modal>
+									<x-slot name="trigger">
+										<div class="text-red-600 hover:text-red-900">
+											<x-icon name="trash" class="w-5 h-5"></x-icon>
+										</div>
+									</x-slot>
+									<x-slot name="title">
+										Conferma eliminazione
+									</x-slot>
+									Sei sicuro di voler eliminare la cartella "{{ $folder->name }}"?
+									<x-slot name="footer">
+										<x-link-button x-on:click="open = false">Annulla</x-link-button>
+										<x-danger-button class="ml-2" wire:click="deleteFolder({{ $folder->id }})"
+										                 wire:loading.attr="disabled">
+											Elimina
+										</x-danger-button>
+									</x-slot>
+								</x-modal>
+							@endcan
 						</div>
 					</x-table.td>
 				</tr>
